+++ lib/models_at.dart
@@ -2,5 +2,14 @@ library models_at;
 
 export 'src/advanced_trade/models/account.dart';
 export 'src/advanced_trade/models/credential.dart';
+export 'src/advanced_trade/models/fill.dart';
+export 'src/advanced_trade/models/order.dart';
+export 'src/advanced_trade/models/order_configuration.dart';
+export 'src/advanced_trade/models/order_configuration/limit_gtc.dart';
+export 'src/advanced_trade/models/order_configuration/limit_gtd.dart';
+export 'src/advanced_trade/models/order_configuration/market_ioc.dart';
+export 'src/advanced_trade/models/order_configuration/stop_limit_gtc.dart';
+export 'src/advanced_trade/models/order_configuration/stop_limit_gtd.dart';
+export 'src/advanced_trade/models/product.dart';
 export 'src/advanced_trade/models/trade.dart';
 export 'src/shared/models/signature.dart';