+++ lib/src/advanced_trade/models/trade.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:coinbase_cloud_exchange_dart_client/src/shared/services/tools.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/shared/services/tools.dart';
 
 class Trade {
   final String? tradeId;