+++ lib/src/advanced_trade/services/signature.dart
@@ -1,4 +1,4 @@
-import 'package:coinbase_cloud_exchange_dart_client/src/shared/models/signature.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/shared/models/signature.dart';
 import 'package:crypto/crypto.dart';
 
 Signature signature(String secret, String httpMethod, String requestPath,