+++ test/pro/rest/accounts_test.dart
@@ -1,8 +1,8 @@
 import 'dart:io' show Platform;
 
-import 'package:coinbase_cloud_exchange_dart_client/src/pro/models/account.dart';
-import 'package:coinbase_cloud_exchange_dart_client/src/pro/models/credential.dart';
-import 'package:coinbase_cloud_exchange_dart_client/src/pro/rest/accounts.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/pro/models/account.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/pro/models/credential.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/pro/rest/accounts.dart';
 import 'package:test/test.dart';
 
 Map<String, String> envVars = Platform.environment;