+++ test/pro/rest/products/ticker_test.dart
@@ -1,7 +1,7 @@
 import 'dart:io' show Platform;
 
-import 'package:coinbase_cloud_exchange_dart_client/src/pro/models/ticker.dart';
-import 'package:coinbase_cloud_exchange_dart_client/src/pro/rest/products/ticker.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/pro/models/ticker.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/pro/rest/products/ticker.dart';
 import 'package:test/test.dart';
 
 Map<String, String> envVars = Platform.environment;