+++ test/pro/services/signature_test.dart
@@ -1,5 +1,5 @@
-import 'package:coinbase_cloud_exchange_dart_client/src/pro/services/signature.dart';
-import 'package:coinbase_cloud_exchange_dart_client/src/shared/models/signature.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/pro/services/signature.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/shared/models/signature.dart';
 import 'package:test/test.dart';
 
 void main() {