+++ test/shared/services/tools_test.dart
@@ -1,4 +1,4 @@
-import 'package:coinbase_cloud_exchange_dart_client/src/shared/services/tools.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/shared/services/tools.dart';
 import 'package:test/test.dart';
 
 void main() {